实现 analogWriteFrequency 和 analogWriteResolution 将map函数由long改为long long 实现analogReadResolution，但目前并不支持ADC调整分辨率 修复没有将avr文件夹下的.c文件加入的问题 Arduino.h包含math.h

+++ core/WMath_c.c
@@ -38,7 +38,7 @@ void randomSeed(unsigned long seed)
     }
 }
 
-long map(long x, long in_min, long in_max, long out_min, long out_max)
+long long map(long long x, long long in_min, long long in_max, long long out_min, long long out_max)
 {
     return (x - in_min) * (out_max - out_min) / (in_max - in_min) + out_min;
 }
